refactor firebase function
