+++ src/services/UsersServices.js
@@ -1,28 +1,31 @@
 import { db } from 'config/firebase/firebase';
-import {
-    getFirestore,
-    query,
-    getDocs,
-    collection,
-    where,
-    addDoc,
-    doc,
-    getDocFromServer,
-    getDoc
-} from 'firebase/firestore';
+import { collection, doc, getDoc, setDoc } from 'firebase/firestore';
 
 const userCollectionRef = collection(db, 'users');
-console.log(userCollectionRef);
-console.log(db);
 class UsersServices {
-    getUser = async (uid) => {
-        const q = query(collection(db, 'users'), where('uid', '==', `${uid}`));
-        //Find user in db with quey
-        const docs = await getDocs(q);
-
-        if (docs.docs.length !== 0) {
-            return docs.docs[0].data();
+    addUser = async (user) => {
+        try {
+            setDoc(doc(db, 'users', user.uid.toString()), {
+                uid: user.uid,
+                name: user.displayName || user.name,
+                authProvider: user.provider,
+                email: user.email,
+                cart: [],
+                phoneNumber: null,
+                address: {
+                    city: null,
+                    district: null,
+                    country: null,
+                    addressNumber: null
+                }
+            });
+        } catch (err) {
+            console.log(err);
         }
+    };
+    getUser = (id) => {
+        const userDoc = doc(db, 'users', id);
+        return getDoc(userDoc);
     };
 }
